requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.28
starlette==0.37.2
psycopg2-binary==2.9.9
bcrypt==4.1.3
mysql-connector-python==9.0.0
uvicorn==0.22.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyYAML==6.0
python-dotenv==1.0.0
google-cloud-storage==2.10.0
email-validator==2.0.0
pydantic-settings==2.6.1
GeoAlchemy2==0.17.1
python-multipart
pydantic
geojson
fastapi-mail